--- src/components/tag/Tag.styles.ts
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const TagWrapper = styled.span`
-  background-color: #add8e6;
-  padding: 4px 8px;
-  border-radius: 16px;
-  font-size: 12px;
-  color: #fff;
-`;
