--- src/components/tag/Tag.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import {
-  TagWrapper,
-} from './Tag.styles';
-
-const Tag: React.FC = () => {
-  return (
-    <TagWrapper>Detail</TagWrapper>
-  );
-};
-
-export default Tag;
